// react-module/src/openapi/models/atomic-product-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductExtendedDto
 */
export interface AtomicProductExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductExtendedDto
     */
    barcodeFromId?: any;
    /**
     * 
     * @type {LayerShape}
     * @memberof AtomicProductExtendedDto
     */
    layerShape?: any;
    /**
     * 
     * @type {AtomicProductDefinitionBasicDto}
     * @memberof AtomicProductExtendedDto
     */
    atomicProductDefinition?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductExtendedDto
     */
    layerStatus?: AtomicProductExtendedDtoLayerStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof AtomicProductExtendedDto
     */
    layerNumber?: any;
    /**
     * 
     * @type {ProductionOrderExtendedDto}
     * @memberof AtomicProductExtendedDto
     */
    productionOrder?: any;
    /**
     * 
     * @type {StorageAreaExtendedDto}
     * @memberof AtomicProductExtendedDto
     */
    storageArea?: any;
    /**
     * 
     * @type {OperationDiagramData}
     * @memberof AtomicProductExtendedDto
     */
    operationDiagramData?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicProductExtendedDtoLayerStatusEnum {
    INITIAL = 'INITIAL',
    SUCCESS = 'SUCCESS',
    BROKEN = 'BROKEN',
    ISPARTOFANOTHER = 'IS_PART_OF_ANOTHER',
    ISSEPARATED = 'IS_SEPARATED',
    INSTOCK = 'IN_STOCK',
    DOESNOTEXIST = 'DOES_NOT_EXIST'
}