// react-module/src/openapi/models/operation-diagram-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface OperationDiagramData
 */
export interface OperationDiagramData {
    /**
     * 
     * @type {number}
     * @memberof OperationDiagramData
     */
    zoomLevel?: any;
    /**
     * 
     * @type {Position}
     * @memberof OperationDiagramData
     */
    offset?: any;
    /**
     * 
     * @type {{ [key, string]: NodeData;}}
     * @memberof OperationDiagramData
     */
    nodeDataMap?: any;
}